+++ src/main/java/com/shoppingcartservice/sachin/Reposistories/Order/OrderRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepo extends JpaRepository<Orders,Integer> {
-    Orders findOrderByOrderId(Integer orderId);
+public interface OrderRepo extends JpaRepository<Orders,Long> {
+    Orders findOrderByOrderId(Long orderId);
     List<Orders> findOrderByUserProfileOrderByOrderIdDesc(UserProfile userProfile);
 }
